Fix multi bucket setups using same credentials (#33)
Currently, multi bucket setups as described in #25 and in the
[carrierwave wiki](https://github.com/carrierwaveuploader/carrierwave/wiki/How-to:-Define-different-storage-configuration-for-each-Uploader.)
are broken. This is because the connection cache does not only cache the
storage connection, but the whole bucket for the first uploader that
gets called.

If you have 2 uploaders with 2 different buckets, but using the same
credentials, only one of them will work correctly. This PR fixes the
issue by only caching the storage connection on the Gcloud class level,
while buckets are cached on the GcloudFile instance level.